filename update
updated file name to be more in line with other formatting, see last commit for what has changed
